requirements.txt
psycopg2-binary
python-dotenv
black
